+++ src/select/type.ts
@@ -2,18 +2,34 @@
 
 /**
  * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
- * updated at 2021-12-12 19:17:30
  * */
 
+import { InputProps } from '../input';
+import { InputValue } from '../input';
 import { PopupProps } from '../popup';
+import { SelectInputProps } from '../select-input';
+import { TagInputProps } from '../tag-input';
+import { TagProps } from '../tag';
+import { SelectInputValueChangeContext } from '../select-input';
+import { PopupVisibleChangeContext } from '../popup';
 import { TNode, SizeEnum } from '../common';
 
 export interface TdSelectProps<T extends SelectOption = SelectOption> {
+  /**
+   * 宽度随内容自适应
+   * @default false
+   */
+  autoWidth?: boolean;
   /**
    * 是否有边框
    * @default true
    */
   bordered?: boolean;
+  /**
+   * 无边框模式
+   * @default false
+   */
+  borderless?: boolean;
   /**
    * 是否可以清空选项
    * @default false
@@ -47,6 +63,18 @@ export interface TdSelectProps<T extends SelectOption = SelectOption> {
    * @default false
    */
   filterable?: boolean;
+  /**
+   * 透传 Input 输入框组件的全部属性
+   */
+  inputProps?: InputProps;
+  /**
+   * 输入框的值
+   */
+  inputValue?: InputValue;
+  /**
+   * 输入框的值，非受控属性
+   */
+  defaultInputValue?: InputValue;
   /**
    * 用来定义 value / label 在 `options` 中对应的字段别名
    */
@@ -82,33 +110,61 @@ export interface TdSelectProps<T extends SelectOption = SelectOption> {
    */
   options?: Array<T>;
   /**
+   * 面板内的底部内容
+   */
+  panelBottomContent?: string | TNode;
+  /**
+   * 面板内的顶部内容
+   */
+  panelTopContent?: string | TNode;
+  /**
    * 占位符
-   * @default ''
    */
   placeholder?: string;
   /**
-   * 透传给 popup 组件的参数
+   * 透传给 popup 组件的全部属性
    */
   popupProps?: PopupProps;
+  /**
+   * 是否显示下拉框
+   */
+  popupVisible?: boolean;
   /**
    * 组件前置图标
    */
   prefixIcon?: TNode;
+  /**
+   * 只读状态，值为真会隐藏输入框，且无法打开下拉框
+   * @default false
+   */
+  readonly?: boolean;
   /**
    * 多选且可搜索时，是否在选中一个选项后保留当前的搜索关键词
    * @default false
    */
   reserveKeyword?: boolean;
   /**
-   * 【讨论中】是否显示全选
-   * @default false
+   * 【开发中】透传 SelectInput 筛选器输入框组件的全部属性
    */
-  showCheckAlll?: boolean;
+  selectInputProps?: SelectInputProps;
+  /**
+   * 是否显示右侧箭头，默认显示
+   * @default true
+   */
+  showArrow?: boolean;
   /**
    * 组件尺寸
    * @default medium
    */
   size?: SizeEnum;
+  /**
+   * 【开发中】透传 TagInput 标签输入框组件的全部属性
+   */
+  tagInputProps?: TagInputProps;
+  /**
+   * 【开发中】透传 Tag 标签组件全部属性
+   */
+  tagProps?: TagProps;
   /**
    * 选中值
    */
@@ -118,9 +174,13 @@ export interface TdSelectProps<T extends SelectOption = SelectOption> {
    */
   defaultValue?: SelectValue;
   /**
+   * 选中值
+   */
+  modelValue?: SelectValue;
+  /**
    * 自定义选中项呈现方式
    */
-  valueDisplay?: TNode<{ value: T[]; onClose: () => void }>;
+  valueDisplay?: string | TNode<{ value: SelectValue; onClose: (index: number, item?: any) => void }>;
   /**
    * 用于控制选中值的类型。假设数据选项为：`[{ label: '姓名', value: 'name' }]`，value 表示值仅返回数据选项中的 value， object 表示值返回全部数据。
    * @default value
@@ -131,9 +191,12 @@ export interface TdSelectProps<T extends SelectOption = SelectOption> {
    */
   onBlur?: (context: { value: SelectValue; e: FocusEvent | KeyboardEvent }) => void;
   /**
-   * 选中值变化时触发
+   * 选中值变化时触发，`context. trigger` 表示触发变化的来源
    */
-  onChange?: (value: SelectValue) => void;
+  onChange?: (
+    value: SelectValue,
+    context: { trigger: SelectValueChangeTrigger; e?: MouseEvent | KeyboardEvent },
+  ) => void;
   /**
    * 点击清除按钮时触发
    */
@@ -150,6 +213,14 @@ export interface TdSelectProps<T extends SelectOption = SelectOption> {
    * 输入框获得焦点时触发
    */
   onFocus?: (context: { value: SelectValue; e: FocusEvent | KeyboardEvent }) => void;
+  /**
+   * 输入框值发生变化时触发，`context.trigger` 表示触发输入框值变化的来源：文本输入触发、清除按钮触发、失去焦点等
+   */
+  onInputChange?: (value: InputValue, context?: SelectInputValueChangeContext) => void;
+  /**
+   * 下拉框显示或隐藏时触发
+   */
+  onPopupVisibleChange?: (visible: boolean, context: PopupVisibleChangeContext) => void;
   /**
    * 多选模式下，选中数据被移除时触发
    */
@@ -208,6 +279,8 @@ export interface SelectKeysType {
 }
 
 export type SelectValue<T extends SelectOption = SelectOption> = string | number | T | Array<SelectValue<T>>;
+
+export type SelectValueChangeTrigger = 'clear' | 'tag-remove' | 'backspace' | 'check' | 'uncheck';
 
 export interface SelectRemoveContext<T> {
   value: string | number;