+++ src/table/hooks/tree-store.ts
@@ -384,11 +384,13 @@ class TableTreeStore<T extends TableRowData = TableRowData> {
         if (!parentExpanded) {
           newData.push(item);
         }
+        this.treeDataMap.set(rowValue, state);
         if (children?.length && !originalExpanded) {
           // 同步更新父元素的展开数量
           let tmpParent = parent;
           while (tmpParent?.row) {
             tmpParent.expandChildrenLength += children.length;
+            this.treeDataMap.set(tmpParent.id, tmpParent);
             tmpParent = tmpParent.parent;
           }
           // 继续子元素
@@ -417,7 +419,7 @@ class TableTreeStore<T extends TableRowData = TableRowData> {
       }
       const children = get(item, keys.childrenKey);
       if (children?.length) {
-        this.expandAll(children, keys);
+        this.foldAll(children, keys);
       }
     }
     return newData;
@@ -463,6 +465,25 @@ class TableTreeStore<T extends TableRowData = TableRowData> {
       treeDataMap.set(rowValue, state);
       if (children?.length) {
         this.initialTreeDataMap(treeDataMap, children, column, keys, level + 1, state);
+      }
+    }
+  }
+
+  // column.checkProps 和 column.disabled 会影响行的禁用状态，因此当列发生变化时，需要重置禁用状态
+  updateDisabledState(dataSource: T[], column: PrimaryTableCol, keys: KeysType) {
+    for (let i = 0, len = dataSource.length; i < len; i++) {
+      const item = dataSource[i];
+      const rowValue = get(item, keys.rowKey);
+      if (rowValue === undefined) {
+        log.error('EnhancedTable', '`rowKey` could be wrong, can not get rowValue from `data` by `rowKey`.');
+        return;
+      }
+      const state = this.treeDataMap.get(rowValue);
+      state.disabled = isRowSelectedDisabled(column, item, i);
+      this.treeDataMap.set(rowValue, state);
+      const children = get(item, keys.childrenKey);
+      if (children?.length) {
+        this.updateDisabledState(children, column, keys);
       }
     }
   }